Conversion function should be in core
