+++ src/poliastro/core/elements.py
@@ -300,3 +300,26 @@ def rv2coe(k, r, v, tol=1e-8):
         nu = np.arctan2(r.dot(cross(h, e)) / norm(h), r.dot(e)) % (2 * np.pi)
 
     return p, ecc, inc, raan, argp, nu
+
+
+@jit
+def mee2coe(p, f, g, h, k, L):
+    """Converts from modified equinoctial orbital elements to classical
+    orbital elements.
+
+    The definition of the modified equinoctial orbital elements is taken from
+    [Walker, 1985].
+
+    Note
+    -----
+    The conversion is always safe because arctan2 works also for 0, 0
+    arguments.
+
+    """
+    ecc = np.sqrt(f ** 2 + g ** 2)
+    inc = 2 * np.arctan(np.sqrt(h ** 2 + k ** 2))
+    lonper = np.arctan2(g, f)
+    raan = np.arctan2(k, h) % (2 * np.pi)
+    argp = (lonper - raan) % (2 * np.pi)
+    nu = (L - lonper) % (2 * np.pi)
+    return p, ecc, inc, raan, argp, nu
